+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Unreleased
 
 * Increase focus contrast of feedback links (PR #731)
+* Add multiple selection option to autocomplete (PR #723)
 
 ## 13.8.1